Fix four memory-safety defects found by an AddressSanitizer run of the test suite
sline.c: the hold-queue timeout callback read entry->token from a freed
entry to build its Debug() lines after sline_release_hold() had freed it.
Copy the token first.

numnicks.c: (ntohs(x) << 16) overflows a signed int for IPv4 addresses
whose first two octets are 192.0 or higher; shift as unsigned.

parse.c/packet.c: parsed message tags point into the sending client's
connection buffer, but the static list send_buffer() falls back to was
only reset when the next line was parsed.  Any send outside a command
handler (exit_client() from a socket EOF/error callback, a timer) used
the previous line's tags -- wrong on its own, and a use-after-free once
that client's connection was gone.  Clear the list after each dispatch.

s_bsd.c: client_timer_callback() exits the client on the TLS handshake
deadline (and read_packet() may exit it too), then dereferenced the freed
struct Client in its trailing assert.  Return early in both cases.

+++ ircd/numnicks.c
@@ -498,7 +498,8 @@ const char* iptobase64(char* buf, const struct irc_in_addr* addr, unsigned int c
 {
   if (irc_in_addr_is_ipv4(addr)) {
     assert(count >= 6);
-    inttobase64(buf, (ntohs(addr->in6_16[6]) << 16) | ntohs(addr->in6_16[7]), 6);
+    inttobase64(buf, ((unsigned int)ntohs(addr->in6_16[6]) << 16)
+                     | ntohs(addr->in6_16[7]), 6);
   } else {
     unsigned int max_start, max_zeros, curr_zeros, zero, ii;
     char *output = buf;
